Adding to commit: Change the item amount from int to float

+++ tilaushallinta/models/tilaus.py
@@ -41,7 +41,7 @@ class Tavara(Base):
     nimi = Column(Text)
 
     hinta = Column(Float)
-    maara = Column(Integer)
+    maara = Column(Float)
 
     tyyppi = Column(Text)
